Unwind some bug compatibility
