+++ safere/src/test/java/org/safere/MatcherTest.java
@@ -279,6 +279,8 @@ class MatcherTest {
     }
   }
 
+
+
   @Nested
   @DisplayName("find()")
   class FindTests {
@@ -1854,6 +1856,28 @@ class MatcherTest {
   @Nested
   @DisplayName("hitEnd()")
   class HitEndTests {
+
+    @Test
+    @DisplayName("hitEnd is true for alternation needing more input (diverges from JDK)")
+    void hitEndTrueForAlternationNeedingMoreInput() {
+      Pattern p = Pattern.compile("abc|abcd");
+      Matcher m = p.matcher("abc");
+      assertThat(m.matches()).isTrue();
+      // SafeRE returns true because it detects that more input ('d') could change the result.
+      // JDK returns false because it stops at the first matching branch.
+      assertThat(m.hitEnd()).isTrue();
+    }
+
+    @Test
+    @DisplayName("hitEnd is true for non-greedy quantifier matches (diverges from JDK)")
+    void hitEndTrueForNonGreedyQuantifierMatches() {
+      Pattern p = Pattern.compile("a*?");
+      Matcher m = p.matcher("a");
+      assertThat(m.matches()).isTrue();
+      // SafeRE returns true because it detects that more input ('a') could change the result.
+      // JDK returns false because it prefers shortest match.
+      assertThat(m.hitEnd()).isTrue();
+    }
 
     @Test
     @DisplayName("hitEnd is true when a variable-length match reaches end")
